server/requirements.txt
flask==2.3.1
python-dotenv==1.0.0
flask-sqlalchemy==3.0.3
flask-jwt-extended==4.4.4
flask-bcrypt==1.0.1